Bugfix: wrong ID number Data-type
Although the Tree.h uses qint64 for the (internal) ID numbers of actions,
aliases, keys, scripts, timers and triggers the rest of the code base uses
an int data-type instead.  This commit changes everything to use "int"
though perhaps it may have been better to explicitly use the Qt qint32 type
that would require careful examination of a LOT of code to check every
location was using the explicit type. As it is this mismatch is not likely
to cause problems as no-one is going to exceed the 2,147,483,647 limit that
using an int type (signed 32-bits) that the likely platforms will use
although strictly speaking the C99 standard only mandates a signed int
maximum value of 32,767 using 16 bits...!

+++ src/ScriptUnit.cpp
@@ -223,7 +223,7 @@ void ScriptUnit::removeScript( TScript * pT )
 }
 
 
-qint64 ScriptUnit::getNewID()
+int ScriptUnit::getNewID()
 {
     return ++mMaxID;
 }
